requirements.txt
watchdog>=3.0.0
cryptography>=41.0.0
uuid6>=2023.5.2
pytest>=7.0.0
pydantic>=2.0.0
pyyaml>=6.0
